requirements.txt
sounddevice
numpy
vosk
pyyaml
requests
piper-tts